Refactor predicates and range from model to logic.
Wait until after PE!
