+++ src/main/java/seedu/dengue/logic/predicates/PersonContainsAgePredicate.java
@@ -1,4 +1,4 @@
-package seedu.dengue.model.predicate;
+package seedu.dengue.logic.predicates;
 
 import java.util.Optional;
