+++ src/main/java/seedu/dengue/logic/predicates/PersonContainsVariantsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.dengue.model.predicate;
+package seedu.dengue.logic.predicates;
 
 import java.util.Set;
 
@@ -18,9 +18,6 @@ public class PersonContainsVariantsPredicate extends PredicateUtil<Person> {
     }
     @Override
     public boolean test(Person person) {
-        if (person.getVariants().containsAll(variants)) {
-            return true;
-        }
-        return false;
+        return person.getVariants().containsAll(variants);
     }
 }
