+++ src/main/java/seedu/dengue/logic/range/Range.java
@@ -1,4 +1,4 @@
-package seedu.dengue.model.range;
+package seedu.dengue.logic.range;
 
 /**
  * Represents a range which can be used to signify age ranges, or date ranges to be used in predicates
