+++ src/main/java/seedu/dengue/logic/range/Start.java
@@ -1,4 +1,4 @@
-package seedu.dengue.model.range;
+package seedu.dengue.logic.range;
 
 import seedu.dengue.model.person.Person;
 
@@ -6,6 +6,6 @@ import seedu.dengue.model.person.Person;
  * Represents the start of a given range
  */
 public interface Start<T> {
-    public boolean isBefore(Person p);
-    public T get();
+    boolean isBefore(Person p);
+    T get();
 }
