+++ src/test/java/seedu/dengue/logic/commands/CommandTestUtil.java
@@ -17,6 +17,12 @@ import java.util.Set;
 
 import seedu.dengue.commons.core.index.Index;
 import seedu.dengue.logic.commands.exceptions.CommandException;
+import seedu.dengue.logic.predicates.FindPredicate;
+import seedu.dengue.logic.range.EndAge;
+import seedu.dengue.logic.range.EndDate;
+import seedu.dengue.logic.range.Range;
+import seedu.dengue.logic.range.StartAge;
+import seedu.dengue.logic.range.StartDate;
 import seedu.dengue.model.DengueHotspotTracker;
 import seedu.dengue.model.Model;
 import seedu.dengue.model.person.Age;
@@ -24,12 +30,6 @@ import seedu.dengue.model.person.Date;
 import seedu.dengue.model.person.Name;
 import seedu.dengue.model.person.Person;
 import seedu.dengue.model.person.SubPostal;
-import seedu.dengue.model.predicate.FindPredicate;
-import seedu.dengue.model.range.EndAge;
-import seedu.dengue.model.range.EndDate;
-import seedu.dengue.model.range.Range;
-import seedu.dengue.model.range.StartAge;
-import seedu.dengue.model.range.StartDate;
 import seedu.dengue.model.variant.Variant;
 import seedu.dengue.testutil.EditPersonDescriptorBuilder;
 
@@ -176,8 +176,8 @@ public class CommandTestUtil {
         Optional<Age> emptyAge = Optional.empty();
         Optional<Date> emptyDate = Optional.empty();
         Set<Variant> emptyVariants = new HashSet<>();
-        Range<Date> emptyDateRange = new Range<Date>(new StartDate(emptyDate), new EndDate(emptyDate));
-        Range<Age> emptyAgeRange = new Range<Age>(new StartAge(emptyAge), new EndAge(emptyAge));
+        Range<Date> emptyDateRange = new Range<>(new StartDate(emptyDate), new EndDate(emptyDate));
+        Range<Age> emptyAgeRange = new Range<>(new StartAge(emptyAge), new EndAge(emptyAge));
         model.updateFilteredPersonList(
                 new FindPredicate(testName, emptySubPostal, emptyAge, emptyDate, emptyVariants,
                         emptyDateRange, emptyAgeRange));
